+++ src/typescript/src/infrastructure/app.ts
@@ -3,7 +3,7 @@ import { dirname } from 'path';
 import fastify from 'fastify';
 import fastifyOpenapiGlue from 'fastify-openapi-glue';
 import Security from './security';
-import { InMemoryLampRepository } from './repositories/in-memory-lamp.repository';
+import { InMemoryLampRepository } from './repositories/InMemoryLampRepository';
 import Service from './services/service';
 
 const __filename = fileURLToPath(import.meta.url);
