+++ src/typescript/src/infrastructure/repositories/InMemoryLampRepository.ts
